config: Add support for expiration policies.
Each backup may now specify rules for when it is okay to remove old backups.
Each rule specifies a time range, and any backup within that range is subject
to the policy of that rule. If multiple rules match, the smallest time interval
wins. There is also a default policy on the top-level config that applies if
the backup doesn't override it.