Print subtotals for -size on selected archives.
The TOTAL line for -size reflects the aggregate storage for the machine key in
question. When the user gives specific archive names or globs, however, it is
also useful to include a subtotal of the storage for just those archives, when
there is more than one.  Add a new SUBTOTAL line in that circumstance.