config: Add support for caching archive listings.
The new config key "list-cache" points to a file where a JSON snapshot of the
archive listing will be kept. This snapshot will be used in preference to a
call out to tarsnap, since listing archives can be very time-consuming and the
archives for a machine should not change much.

The cache can become invalid if you use tarsnap separately to create or delete
archives.