config: Handle list cache invalidation automatically.
Update to a new version of the tarsnap library that knows how to find the cache
sequence number. Store that in the cache along with the listing. Check the hot
value of the sequence number to validate the stored data.

It is still possible for snapback to race with another tarsnap invocation,
which might update the sequence number after we read it but before we have
loaded the cache, but the result will only be stale until the next call, and
will still be consistent.