config: add a manual flag to the backup set.
When this flag is set true, the backup set will not be executed when creating
backups without an explicit list of sets. For example, given sets A, B, C and B
is marked manual, running "snapback" will back up A and C only. To back up B in
this case, use "snapback -c B".