config: Fix a path matching bug due to premature glob expansion
Previously, include globs were expanded eagerly during config parsing.  This is
fine when creating backups, since the backup process will fail anyway if the
glob expansion does not produce any results.

However: When resolving paths (-find) and restoring (-restore), this is not
what we want. The candidate paths are provided by the user, not the filesystem,
and in fact the local filesystem might not even contain the original directory.

To fix this, defer glob expansion until after we know what operation we are
doing, via a now-exported method of the Backup type.